requirements.txt
Django==1.5.5
Pygments==1.6
South==0.8.3
Unipath==1.0
argparse==1.2.1
dj-database-url==0.2.2
dj-static==0.0.5
django-filter==0.7
django-rest-swagger==0.1.11
django-toolbelt==0.0.1
djangorestframework==2.3.8
gunicorn==18.0
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
factory_boy==2.4.1
pytz==2014.4
